requirements.txt
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
pdf2image>=1.16.3
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=0.27.0
qwen-vl-utils>=0.0.1
sentencepiece>=0.1.99
protobuf>=3.20.0

# System dependencies (install separately):
# - poppler-utils (for pdf2image)

